Refactor: 홈 화면의 챗봇 form을 클라이언트 컴포넌트로 분리 후 홈 화면을 SSR로 전환
